Move LoginCommand to more meaningful namespace Add IImageEditor Sync namespaces. Check if blob was deleted before removing record from database Move ImageEditor in Functions to Functions\Services. Change IJAAADbContext to FileRepository in Functions. Remove CustomTestStartum from Functions.UnitTests. Add ImageEditorUnitTests. Update MiniaturizeUnitTests

+++ Src/Core/Application/Images/Commands/DeleteImage/DeleteImageCommand.cs
@@ -38,12 +38,13 @@ namespace Application.Images.Commands.DeleteImage
                     var statusCode = await _blobManagerService.DeleteBlobAsync(blob.Name, cancellationToken);
                     var file = await _fileRepository.GetByNameAsync(blob.Name, cancellationToken);
 
-                    if(file != null)
+                    if (statusCode != HttpStatusCode.Accepted) throw new OperationFailedException(HttpStatusCode.Accepted,
+                        statusCode, nameof(DeleteImageCommandHandler));
+
+                    if (file != null)
                     {
                         await _fileRepository.RemoveAsync(file.Filename, cancellationToken);
                     }
-                    if (statusCode != HttpStatusCode.Accepted) throw new OperationFailedException(HttpStatusCode.Accepted,
-                        statusCode, nameof(DeleteImageCommandHandler));
                 }
 
                 return Unit.Value;
